// src/store/types.ts
export interface AppState {
  loading: boolean
}

export type TeamData = {
  name: string
  logo: string
  price: number
  kudosToTire1: number
  kudosToTire2: number
  kudosToTire3: number
  kudosToTire4: number
  memberLevelName1: string
  memberLevelName2: string
  memberLevelName3: string
  memberLevelName4: string
}